+++ app/endpoints/sillyusers/gen_user.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 import silly
 import uuid
 
@@ -44,9 +45,9 @@ def user_info(user: str = None, id: str = None):
     return result
 
 
-if __name__ == "__main__":
-    id = "123"
-    user = None
-    for i in range(0, 1):
-        x = user_info(user, id)
-        print(x)
+# if __name__ == "__main__":
+#     id = "123"
+#     user = None
+#     for i in range(0, 1):
+#         x = user_info(user, id)
+#         print(x)