+++ app/endpoints/users/models.py
@@ -1,5 +1,6 @@
+# -*- coding: utf-8 -*-
 from typing import List, Optional, Union, Set
-from pydantic import BaseModel, Schema, Json, UUID1
+from pydantic import BaseModel, Schema, Json, UUID1, SecretStr
 import uuid
 from datetime import datetime, timedelta
 import json
@@ -7,6 +8,7 @@ import json
 # Shared properties
 class UserBase(BaseModel):
     # userId: str
+    user_name: str
     firstName: str
     lastName: str
     password: str
@@ -26,11 +28,13 @@ class UserBase(BaseModel):
 
 class UserBaseInDB(UserBase):
     userId: str = None
+    user_name: str
 
 
 # Properties to receive via API on creation
 class UserCreate(UserBase):
     # userId: str
+    user_name: str
     firstName: str
     lastName: str
     password: str
@@ -49,6 +53,11 @@ class UserCreate(UserBase):
 
 class UserBaseInDB(UserBase):
     userId: str = None
+
+
+class UserPwd(UserBase):
+    user_name: str
+    password: SecretStr
 
 
 # Properties to receive via API on update
@@ -74,11 +83,18 @@ class UserUpdate(UserBaseInDB):
     # isActive: bool = True
 
 
-# # Additional properties to return via API
-# class User(UserBaseInDB):
-#     pass
-
-
-# # Additional properties stored in DB
-# class UserInDB(UserBaseInDB):
-#     hashed_password: str
+class UserList(UserBaseInDB):
+    firstName: str
+    lastName: str
+    # password: str
+    title: Optional[str] = None
+    company: Optional[str] = None
+    address: Optional[str] = None
+    city: Optional[str] = None
+    country: Optional[str] = None
+    postal: Optional[str] = None
+    email: Optional[str] = None
+    website: Optional[str] = None
+    description: Optional[str] = None
+    # dateCreate: datetime = None
+    isActive: bool = True